\name{shinyCpp-package}
\alias{shinyCpp-package}
\alias{shinyCpp}
\docType{package}
\title{
\packageTitle{shinyCpp}
}
\description{
\packageDescription{shinyCpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinyCpp}
\packageIndices{shinyCpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shinyCpp}

Maintainer: \packageMaintainer{shinyCpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
